Finished day 10 final project

+++ Day 10/Calculator/Finalalized_calculator.py
@@ -15,13 +15,13 @@ operations = {
 }
 def calculator():
   print(logo) 
-  num1 = int(input("What is the first number?: "))
+  num1 = float(input("What is the first number?: "))
   for symbol in operations:
       print(symbol)
   continue_operation = True
   while continue_operation:
     choosen_symbol = input("Please Choose operation: ")
-    num2 = int(input("What is the second number?: "))
+    num2 = float(input("What is the second number?: "))
     function_operation = operations[choosen_symbol]
     answer = function_operation(num1, num2)
     print("{} {} {} ={}".format(num1, choosen_symbol, num2, answer))
